:recycle: split chopper_logs into ChopperErrorLog, ChopperRequestLog, and ChopperResponseLog classes for enhanced readability
